+++ src/Moq/SequenceExtensions.cs
@@ -48,6 +48,22 @@ namespace Moq
 		}
 
 		/// <summary>
+		/// Return a sequence of tasks, once per call.
+		/// </summary>
+		public static ISetupSequentialResult<Task> PassAsync(this ISetupSequentialResult<Task> setup)
+		{
+			return setup.Returns(() => Task.FromResult(0));
+		}
+
+		/// <summary>
+		/// Return a sequence of tasks, once per call.
+		/// </summary>
+		public static ISetupSequentialResult<ValueTask> PassAsync(this ISetupSequentialResult<ValueTask> setup)
+		{
+			return setup.Returns(() => new ValueTask());
+		}
+
+		/// <summary>
 		/// Throws a sequence of exceptions, once per call.
 		/// </summary>
 		public static ISetupSequentialResult<Task<TResult>> ThrowsAsync<TResult>(this ISetupSequentialResult<Task<TResult>> setup, Exception exception)
@@ -70,6 +86,32 @@ namespace Moq
 				var tcs = new TaskCompletionSource<TResult>();
 				tcs.SetException(exception);
 				return new ValueTask<TResult>(tcs.Task);
+			});
+		}
+
+		/// <summary>
+		/// Throws a sequence of exceptions, once per call.
+		/// </summary>
+		public static ISetupSequentialResult<Task> ThrowsAsync(this ISetupSequentialResult<Task> setup, Exception exception)
+		{
+			return setup.Returns(() =>
+			{
+				var tcs = new TaskCompletionSource<object>();
+				tcs.SetException(exception);
+				return tcs.Task;
+			});
+		}
+
+		/// <summary>
+		/// Throws a sequence of exceptions, once per call.
+		/// </summary>
+		public static ISetupSequentialResult<ValueTask> ThrowsAsync(this ISetupSequentialResult<ValueTask> setup, Exception exception)
+		{
+			return setup.Returns(() =>
+			{
+				var tcs = new TaskCompletionSource<object>();
+				tcs.SetException(exception);
+				return new ValueTask(tcs.Task);
 			});
 		}
 	}